backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
pydantic-settings
python-multipart
passlib[bcrypt]
bcrypt<5.0.0
python-jose[cryptography]
python-dateutil
